random_tests/requirements.txt
discord.py
requests
ai4free
# https://github.com/Kritik-helpingai/Helpingai_T2
websocket
click
g4f
selenium
